resemble_service/pyproject.toml
[project]
name = "coding-challenge"
version = "0.1.0"
description = "Coding challenge"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
 "httpx>=0.28.1",
 "mcp[cli]>=1.3.0",
 "python-dotenv>=1.0.1",
 "resemble>=1.5.0",
]
[[project.authors]]
name = "Tudor Filimon"
email = "tfilimon123@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
coding-challenge = "coding_challenge:main"
